db changes to implement prepared statement queries (prevent sql injection and character escaping)

+++ php/admin/action/post/deletePost.php
@@ -8,22 +8,23 @@ if (isEmpty($id)) {
 
 try {
     $post2delete = PostHandler::getPostByID($id);
-    if ($post2delete->getState() == PostStatus::PUBLISHED) {
-        addInfoMessage("Post '" . $post2delete->getTitle() . "' is published and cannot be deleted");
-    } else {
-        $deletePostRes = PostHandler::deletePost($id);
-
-        if ($deletePostRes == null || !$deletePostRes) {
-            addErrorMessage("Post failed to be deleted");
+    if (isNotEmpty($post2delete)) {
+        if ($post2delete->getState() == PostStatus::PUBLISHED) {
+            addInfoMessage("Post '" . $post2delete->getTitle() . "' is published and cannot be deleted");
         } else {
-            addSuccessMessage("Post successfully deleted");
+            $deletePostRes = PostHandler::deletePost($id);
+            if ($deletePostRes == null || !$deletePostRes) {
+                addErrorMessage("Post failed to be deleted");
+            } else {
+                addSuccessMessage("Post successfully deleted");
+            }
         }
+    } else {
+        addInfoMessage(ErrorMessages::WENT_WRONG);
     }
 } catch (SystemException $ex) {
     logError($ex);
     addErrorMessage(ErrorMessages::GENERIC_ERROR);
-    // you can exit or die here if you prefer - also you can log your error,
-    // or any other steps you wish to take
 }
 
 Redirect(getAdminRequestUri() . "posts");
